ensure phasOnlyOneTokenOfCurrencySymbol behaves
In the case of having a value structures as follows:

`[(s, [(a, 2), (b, -1)])]`, where `s` is a symbol and `a` and `b` are
token names, previously `phasOnlyOneTokenOfCurrencySymbol` would return
`PTrue`. This is incorrect, because the value contains multiple tokens.

Now, using `psymbolValueOf'`, we require that the value be _strictly_
positive, as well as being a single token.

Thanks @colll78!